Rename hash_parser::backend to hash_parser::parser
