+++ compiler/hash-parser/src/parser.rs
