+++ tests/parser/tests/test_runner.rs
@@ -1,8 +1,8 @@
 #![cfg(test)]
 
 use hash_alloc::Castle;
-use hash_parser::backend::HashParser;
-use hash_pipeline::{Parser, Sources, Module};
+use hash_parser::parser::HashParser;
+use hash_pipeline::{Module, Parser, Sources};
 use hash_source::SourceId;
 use hash_utils::testing::TestingInput;
 use hash_utils_testing_macros::generate_tests;
